# Application Programming Interface/BuildingBug.API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BuildingBugAPI",
    "version": "v1"
  },
  "paths": {
...
    "/api/v{version}/Deslocacao/new": {
      "post": {
        "tags": [
          "Deslocacao"
        ],
        "summary": "Marcar uma Deslocação",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateDeslocacaoRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Deslocacao/update": {
      "put": {
        "tags": [
          "Deslocacao"
        ],
        "summary": "Atualizar uma deslocação",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateDeslocacaoRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Deslocacao/delete/{id}": {
      "delete": {
        "tags": [
          "Deslocacao"
        ],
        "summary": "Eliminar uma deslocação",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Deslocacao/{id}": {
      "get": {
        "tags": [
          "Deslocacao"
        ],
        "summary": "Obter uma deslocação",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Deslocacao/all": {
      "get": {
        "tags": [
          "Deslocacao"
        ],
        "summary": "Obter todas as deslocações",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/EstadoServico/new": {
      "post": {
        "tags": [
          "EstadoServico"
        ],
        "summary": "Criar um estado de serviço",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateEstadoServicoRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/EstadoServico/update": {
      "put": {
        "tags": [
          "EstadoServico"
        ],
        "summary": "Atualizar um estado de serviço",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateEstadoServicoRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/EstadoServico/delete/{id}": {
      "delete": {
        "tags": [
          "EstadoServico"
        ],
        "summary": "Eliminar um estado de serviço",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/EstadoServico/{id}": {
      "get": {
        "tags": [
          "EstadoServico"
        ],
        "summary": "Obter o estado de serviço",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/EstadoServico/all": {
      "get": {
        "tags": [
          "EstadoServico"
        ],
        "summary": "Obter todos os estados de serviço",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v{version}/Servico/new": {
      "post": {
        "tags": [
          "Servico"
        ],
        "summary": "Pedir serviço",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateServicoRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/update": {
      "put": {
        "tags": [
          "Servico"
        ],
        "summary": "Atualizar serviço",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateServicoRequestContract"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/delete/{id}": {
      "delete": {
        "tags": [
          "Servico"
        ],
        "summary": "Eliminar serviço",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/{id}": {
      "get": {
        "tags": [
          "Servico"
        ],
        "summary": "Obter serviço",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Servico/all": {
      "get": {
        "tags": [
          "Servico"
        ],
        "summary": "Obter todos os serviços",
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateDeslocacaoRequestContract": {
      "type": "object",
      "properties": {
        "data": {
          "format": "date-time",
          "type": "string"
        },
        "servico_id": {
          "format": "int32",
          "type": "integer"
        },
        "detalhes": {
          "type": "string"
        }
      }
    },
    "CreateEstadoServicoRequestContract": {
      "type": "object",
      "properties": {
        "descricao": {
          "type": "string"
        }
      }
    },
...
    "CreateServicoRequestContract": {
      "required": [
        "clienteId",
        "codigoPostalCP",
        "fator",
        "morada",
        "tipoId"
      ],
      "type": "object",
      "properties": {
        "morada": {
          "minLength": 1,
          "type": "string"
        },
        "fator": {
          "format": "int32",
          "type": "integer"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "codigoPostalCP": {
          "format": "int32",
          "type": "integer"
        },
        "tipoId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "UpdateDeslocacaoRequestContract": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "data": {
          "format": "date-time",
          "type": "string"
        },
        "servico_id": {
          "format": "int32",
          "type": "integer"
        },
        "detalhes": {
          "type": "string"
        }
      }
    },
    "UpdateEstadoServicoRequestContract": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "descricao": {
          "type": "string"
        }
      }
    },
...
    "UpdateServicoRequestContract": {
      "required": [
        "avaliacao",
        "clienteId",
        "codigoPostalCP",
        "custo",
        "dataFim",
        "dataInicio",
        "estadoServicoId",
        "fator",
        "id",
        "morada",
        "profissionalId",
        "tipoId"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "morada": {
          "minLength": 1,
          "type": "string"
        },
        "dataInicio": {
          "format": "date-time",
          "type": "string"
        },
        "dataFim": {
          "format": "date-time",
          "type": "string"
        },
        "avaliacao": {
          "format": "int32",
          "type": "integer"
        },
        "custo": {
          "format": "float",
          "type": "number"
        },
        "fator": {
          "format": "int32",
          "type": "integer"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "estadoServicoId": {
          "format": "int32",
          "type": "integer"
        },
        "codigoPostalCP": {
          "format": "int32",
          "type": "integer"
        },
        "profissionalId": {
          "format": "int32",
          "type": "integer"
        },
        "tipoId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insert Token as: 'Bearer your_token_here'"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}